Add files via upload
This contains the convective heat equation in its numerical form and uses CUDA through Cupy. It is important to recognize the way the data is stored in the for loop as a derivative of a list, this is to keep the memory isolated to the GPU and keep it from calling back to the CPU every iteration. As of now, CUDA is around 20-100 times faster for calculations of this PDE. I still need to do analysis on it to ensure it reflects the PDE correctly.